Use exit code of tests (#83)
* Use exit code of tests

Previously, the exit code of `pdm test` was the exit code from `docker
compose down`. This change instead uses the exit code of running the
test.

* Add missing keys to test data

2 tests fail due to authorization not being turned off for tests.

* Drop tables before running rest of test setup

This should make sure we start with a clean slate before we insert our test data.

As a side benefit, it may make tests run faster, since we no longer need
to destroy the postgres container.