Add batch queue endpoint and extract common logic
Add a `GET /batch/job` endpoint which is similar to `POST /job` except
this endpoint gets multiple jobs at once. The amount of jobs to get is
specified by the `n_jobs` query string parameter. The conditions for
distributing a job here is the same as that in `POST /job`. Due to this,
the common logic has been moved into a function `fetch_job` which
contains the appropriate SQLAlchemy query for getting jobs. This
function accepts `n_jobs` to specify how many jobs to fetch from the
database. `POST /job` will simply pass `1` in for this argument, while
the new batch endpoint will pass in whatever is in the query string.
Similarly, a `_build_response` private method has been added that
abstracts away the building of the HTTP response body (which also
happens to be common to both the single and multi-request endpoints).