Add inspector URL to test data
Add inspector URL to `first_safe_second_malicious.json` to records that have a score
greater than 0 (since packages that match rules *must* have an inspector
URL). This was done because of the conditional in L34-35 of
`src/endpoints/scans.py`:
```py
if scan.inspector_url is None:
    continue
```
The previous conditional that checks the score should ensure that the
inspector_url is never actually null in a production environment where
we get real data instead of our own fed in from test data. However,
since we use one big table for everything, most of our columns are
nullable and and "runtime checked." This means that even though the
`inspector_url` will always not be null if the score is greater than
0, the type-checker doesn't know this. So this is mostly a check to
appease the typechecker.