Add ORM models and database migrations
Add SQLAlchemy model for `APIClient` which represents a user in the
database. It has a unique client_id, the hashed argon2 client_secret
(this is the hashed_secret column), a unique username meant as a
human-friendly way to identify a user, and a boolean column `admin` that
indicates if this user has administrator privileges or not.

The database migration has also been updated to reflect the new added
model.