Add dependencies
Add two authentication-related dependencies to dependencies.py:
`check_admin` which is intended for use in FastAPI's path operation
decorator. It will confirm that the current user has the `admin`
privilege. If they do not, it will throw an HTTPException with status
code 403.
The `get_current_client` dependency will return the `APIClient`
SQLAlchemy model for the user associated with the provided
`access_token`.