Make check_admin depend on get_current_client
Since FastAPI allows dependencies to have sub-dependencies themselves,
the `check_admin` dependency should depend on `get_current_client`.
These two functions shared some logic, and it is only reasonable that in
order to check if a client is an administrator, we first need to get
that client (using `get_current_client`).