Remove redundant check in send_report_email function
Remove the redundant check in send_report_email function of
`src/mainframe/endpoints/report.py` that checked that a user wasn't
attempting to report a package that matched no rules without providing
an additional description. However, this check was already being done in
the FastAPI endpoint callback before calling this function, so this if
statement would never be called.