Make GET /rules use rules dependency
Now that we have a FastAPI dependenecy for getting the rules commit hash
and the rules themselves, this endpoint has been modified to have
FastAPI inject that rules dependency into the callback. This was done
in order to more easily call the endpoint function directly in tests,
without having to mock the entire request object.
Note:The dependenecy function had to be renamed using `as` since it
was clashing with the name of the endpoint function.