Bring test coverage up to 100% (#161)
* Remove rule_names association proxy

Remove the rule_names association proxy in SQLAlchemy `Scan` model
because it's not used anywhere in our code currently.

* Remove redundant check in send_report_email function

Remove the redundant check in send_report_email function of
`src/mainframe/endpoints/report.py` that checked that a user wasn't
attempting to report a package that matched no rules without providing
an additional description. However, this check was already being done in
the FastAPI endpoint callback before calling this function, so this if
statement would never be called.

* Make GET /rules use rules dependency

Now that we have a FastAPI dependenecy for getting the rules commit hash
and the rules themselves, this endpoint has been modified to have
FastAPI inject that rules dependency into the callback. This was done
in order to more easily call the endpoint function directly in tests,
without having to mock the entire request object.
Note:The dependenecy function had to be renamed using `as` since it
was clashing with the name of the endpoint function.

* Rewrite rules.py to be more testable

* Add tests to get coverage up to 100%