Remove "VERSION" from GitHub Accept header
Most of the blogposts online show using "application/vnd.github.VERSION.sha" to retrieve just the SHA, but according to GitHub's [docs](https://docs.github.com/en/rest/using-the-rest-api/media-types?apiVersion=2022-11-28#sha-media-type-for-commits-commit-comparison-and-pull-requests) it's supposed to be "application/vnd.github.sha".

I have tested both, and they appear to work the same. I don't know what the difference is, but I'd like to use the one that's officially documented and supported.

Signed-off-by: Bradley Reynolds <bradley.reynolds@darbia.dev>