Modify existing tests to work with job cache
Remove the test that tests if the API handles accordingly if a client
tries to submit results for a package which we don't have records for.
This needs to be done because the cache batches results, so we don't get
immediate feedback on if the submitted result was valid or not.

The function signatures for the endpoints have also changed (they accept the
job_cache dependency) so the tests will have to provide that.