Use sessionmaker in `JobCache`
The `JobCache` class saves a `Session` instance as part of it's state. However,
SQLAlchemy's sessions are not designed to be thread safe.
To fix this, I've instead saved a session *factory* as part of the class's cache
so that each method can create a fresh session as it needs to.