Add option to enable JSON logging
Add an option in `constants.py` to enable JSON logging. This will be set
to False by default, in the philosophy that future settings similar to
this will have development-centric defaults.

We would rather have pretty color logging work out of the box in
development, and go through the trouble of adding a few flags in
production, than have the developer have to add those options in every
time locally.