Add `db` pytest marker
This marker should be used on all tests that reach for a real database
to denote that they may be slow and require a database container to be
spun up before this test can be run. This will allow developers who are
not making any database changes to run their tests very quickly. All
tests (including database ones) should still be run in CI.