Fix typing errors
Pyright can't infer that lookup_package_info returns a Page if `page`
and `size` are passed in, and returns a Sequence otherwise.
Using an isinstance check instead of checking `page` and `size` gets us
partway, however the generic type information for Page[T] is lost within
this type guard, so we must still ignore type checking a bit to prevent
"varaiable is of unknown type" errors.